# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio API",
    "version": "1.2.0",
    "description": "API for contact submissions, health, admin session flows, message retrieval, metrics, and telemetry ingestion."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Get service health",
        "responses": {
          "200": {
            "description": "Health payload"
          }
        }
      }
    },
    "/api/contact": {
      "post": {
        "summary": "Submit contact message",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "subject", "message"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "subject": { "type": "string" },
                  "message": { "type": "string" },
                  "website": { "type": "string", "description": "Honeypot field" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Accepted" },
          "400": { "description": "Validation error" },
          "429": { "description": "Rate limited" }
        }
      }
    },
    "/api/messages": {
      "get": {
        "summary": "Get admin messages",
        "security": [{ "basicAuth": [] }],
        "responses": {
          "200": { "description": "Message list" },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Locked out" }
        }
      }
    },
    "/api/admin/session": {
      "get": {
        "summary": "Get admin session state",
        "responses": {
          "200": { "description": "Session state" }
        }
      }
    },
    "/api/admin/login": {
      "post": {
        "summary": "Start admin session",
        "security": [{ "basicAuth": [] }],
        "responses": {
          "200": { "description": "Session started" },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Locked out" }
        }
      }
    },
    "/api/admin/logout": {
      "post": {
        "summary": "Close admin session",
        "responses": {
          "200": { "description": "Session closed" },
          "403": { "description": "Invalid CSRF token" }
        }
      }
    },
    "/api/admin/metrics": {
      "get": {
        "summary": "Get admin diagnostics metrics",
        "responses": {
          "200": { "description": "Metrics payload" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/telemetry": {
      "post": {
        "summary": "Accept privacy-friendly telemetry events",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["event", "path"],
                "properties": {
                  "event": { "type": "string" },
                  "path": { "type": "string" },
                  "locale": { "type": "string" }
                }
              }
            },
            "text/plain": {
              "schema": {
                "type": "string",
                "description": "JSON payload string"
              }
            }
          }
        },
        "responses": {
          "202": { "description": "Accepted" },
          "400": { "description": "Invalid payload" }
        }
      }
    },
    "/api/openapi.json": {
      "get": {
        "summary": "Get OpenAPI specification",
        "responses": {
          "200": { "description": "Spec JSON" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}